Переработано разделение ролей, доабвлен автоматический вход в систему после регистрации

+++ spring/demo/src/main/java/com/example/demo/controller/LessonController.java
@@ -3,6 +3,7 @@ package com.example.demo.controller;
 import com.example.demo.domain.Course;
 import com.example.demo.domain.Lesson;
 import com.example.demo.dto.LessonDto;
+import com.example.demo.exception.AccessDeniedException;
 import com.example.demo.exception.NotFoundException;
 import com.example.demo.service.CourseService;
 import com.example.demo.service.LessonService;
@@ -24,6 +25,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletRequest;
+
 @Controller
 @RequestMapping({"/lesson"})
 public class LessonController {
@@ -41,7 +44,11 @@ public class LessonController {
     }
 
     @GetMapping({"/new"})
-    public String lessonForm(Model model, @RequestParam("course_id") long courseId) {
+    public String lessonForm(Model model, @RequestParam("course_id") long courseId,
+                             HttpServletRequest request) {
+        if (!request.isUserInRole("ROLE_ADMIN")) {
+            throw new AccessDeniedException();
+        }
         model.addAttribute("courseId", courseId);
         model.addAttribute("lesson", new LessonDto(courseId));
         return "lesson_form";
@@ -54,11 +61,14 @@ public class LessonController {
         return "lesson_form";
     }
 
-    @PostMapping(
-            params = {"submit"}
-    )
-    public String submitLessonForm(LessonDto lessonDto, Model model, BindingResult bindingResult) {
+    @PostMapping(params = {"submit"})
+    public String submitLessonForm(LessonDto lessonDto, Model model,
+                                   BindingResult bindingResult,
+                                   HttpServletRequest request) {
         this.statisticsCounter.countHandlerCall("/lesson/submit");
+        if (!request.isUserInRole("ROLE_ADMIN")) {
+            throw new AccessDeniedException();
+        }
         if (bindingResult.hasErrors()) {
             return "lesson_form";
         } else {
@@ -73,21 +83,16 @@ public class LessonController {
     }
 
     @DeleteMapping({"/{id}"})
-    public String deleteLesson(Model model, @PathVariable("id") Long id) {
+    public String deleteLesson(Model model, @PathVariable("id") Long id,HttpServletRequest request) {
         this.statisticsCounter.countHandlerCall("/course/{id} - delete");
+        if (!request.isUserInRole("ROLE_ADMIN")) {
+            throw new AccessDeniedException();
+        }
         Course course = this.lessonService.lessonById(id).getCourse();
         this.lessonService.deleteLesson(id);
         model.addAttribute("course", CourseMapper.mapLessonToDto(course));
         model.addAttribute("lessons", this.lessonService.lessonsWithoutText(course.getId()));
         model.addAttribute("users", course.getUsers());
         return "course_form";
-    }
-
-    @ExceptionHandler({NotFoundException.class})
-    public ModelAndView notFoundExceptionHandler() {
-        this.statisticsCounter.countHandlerCall("/course_not_found");
-        ModelAndView modelAndView = new ModelAndView("not_found");
-        modelAndView.setStatus(HttpStatus.NOT_FOUND);
-        return modelAndView;
     }
 }
